requirements.txt
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
# psycopg2-binary
psycopg2==2.9.3
whitenoise==6.2.0